refactor(maestro): move the declaration-keyword hover tests into their own module
The file crossed the 350-line source-length limit; the tests move to a child
tests module, the convention used across the crate.
